+++ android/app/src/main/java/com/remriel/ovotimer/MainActivity.java
@@ -1,5 +1,12 @@
 package com.remriel.ovotimer;
 
+import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
 
-public class MainActivity extends BridgeActivity {}
+public class MainActivity extends BridgeActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(OvoAlarmPlugin.class);
+        super.onCreate(savedInstanceState);
+    }
+}